fix: fixed api calls in dashboard
