+++ webapp/src/app/core/core.module.ts
@@ -31,8 +31,6 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { RequestInterceptorService } from './services/request-interceptor.service';
 import { AuthSessionStorageService } from './services/auth-session-storage.service';
 import {RecentlyViewedObservableService} from './services/recently-viewed-observable.service';
-import { TableStateService } from './services/table-state-service.service';
-
 @NgModule({
   imports: [
     CommonModule,
@@ -55,7 +53,6 @@ import { TableStateService } from './services/table-state-service.service';
     AdalService,
     AuthSessionStorageService,
     RecentlyViewedObservableService,
-    TableStateService,
     {
       provide: HTTP_INTERCEPTORS,
       useClass: RequestInterceptorService,
